fix(l10n): Move Intl.direction to Intl reducer
