+++ common/reducers/Intl.js
@@ -1,6 +1,10 @@
-const {DEFAULT_LOCALE} = require("common/constants");
+const {DEFAULT_LOCALE, RTL_LIST} = require("common/constants");
 const INITIAL_STATE = {locale: DEFAULT_LOCALE, strings: {}, direction: "ltr"};
 const getLocalizedStrings = require("common/getLocalizedStrings");
+
+function getDirection(locale) {
+  return (RTL_LIST.indexOf(locale.split("-")[0]) >= 0) ? "rtl" : "ltr";
+}
 
 function Intl(prevState = INITIAL_STATE, action) {
   switch (action.type) {
@@ -8,16 +12,17 @@ function Intl(prevState = INITIAL_STATE, action) {
       if (!action.data) {
         return prevState;
       }
-      return Object.assign({}, prevState, {
-        locale: action.data.locale,
-        strings: getLocalizedStrings(action.data.locale),
-        direction: action.data.direction
-      });
+      return {
+        locale: action.data,
+        strings: getLocalizedStrings(action.data),
+        direction: getDirection(action.data)
+      };
     default:
       return prevState;
   }
 }
 
 Intl.INITIAL_STATE = INITIAL_STATE;
+Intl.getDirection = getDirection;
 
 module.exports = Intl;
